requirements.txt
flask
flask-restx
ansible-runner
werkzeug
pyyaml
pytest